check_for_updates as opt-in
This patchset changes the default update procedures to an opt in instead of a opt out.
It reverses the defaults for check_for_updates and for bare_update_check
to don't check for updates per default and if user enables it in nagios.cfg
it won't send further update informations to the update server until the user
decide to do so.