Fix plugins configure option to --with-nagios-user
The configure option of the NagiosPlugins should be --with-nagios-user instead of --with-icinga-user. There is no --with-icinga-user option in the NagiosPlugins' configure script.
I've updated the resulting html file as well.

This patch is similar to commit 3fa9c8c4904726c2df2bea7159b485e2de8064ce, fixing the quickstart guide without using idoutils.