add service_check_timeout_state	configuration variable (Bill McGonigle)
This patch adds a variable called 'service_check_timeout_state'
which allows the admin to define the state that is returned
when a service check times out.

The #ifdef SERVICE_CHECK_TIMEOUTS_RETURN_UNKNOWN was available
for re-compiling, which works, but then those users are unable
to keep up with their distro's updates and it may be beyond
the skill of many.

This patch moves that idea into a variable, allows any of four
states to be chosen ('critical' remaining the default), and does
away with the #ifdef (which should be obsolete now).

refs #275