add in sync retention facility (Ton Voon, Opsera Ltd)
this allows to read saved sync.dat from master into
the slaves and update the status in case. this is
mainly meant for distributed systems and only affects
the core.

the automated generation of sync.dat is being done
using seperated scripting from opsview.
it's more or less OPT-IN, but we're gonna take it -
maybe it is possible to get opsview working with
icinga though.

    * stop icinga
    * get a host from retentation.dat and copy it to a new file named sync.dat
    * modify last_check into the future, and modify a param, e.g. is_flapping=1
    * start icinga
    * check if sync.dat gets removed
    * check the web for an update
    * stop icinga and check retention.dat for update

works as expected - retention.dat gets updated.

based on this feature people can sync their slaves with
the current master state, in case this is needed in a
distributed setup.

credits to ton voon and opsera.

fixes #412