remove dead code (Andreas Ericsson)
Remove some dead code
my_free() has been a macro since 2007-10-22, but the function
version of it has been hanging around (commented out) quite
long enough, so remove it.

Remove more dead code
None of the code removed has been in use for over a year.
It's easy to pull it out of the scm in case we need it
later, so let's just remove it entirely.

refs #388