create logging.h, with stubs for the common function in cgiutils.c (Andreas Ericsson)
[--snip--]
The (extremely) many '#ifdef NSCORE' entries scattered throughout
the shared code makes making major changes a real headache. This
is a first step to let us work the code without worrying if we're
the cgi's or the daemon.

The fact that we're removing almost 500 lines more than we're
adding is a sure sign it's a step in the right direction.
[--snip--]

fits in Icinga too, applied.

fixes #394