xodtemplate: Use macro magic to get rid of 900 LoC (Andreas Ericsson)
[--snip--]
Besides being extremely boring lines of code, changing all of
them would be both a total nightmare and extremely error-prone.

The macro isn't the simplest of simple to follow, as it relies
on the preprocessor to concatenate preprocessor tokens to make
it real code, but it works well enough and compiles nicely on
my truly ancient system with gcc 2.95, so it's not exactly
stretching for the most modern of features.
[--snip--]

this one's really great, fits perfectly into Icinga :)

fixes #408