base/commands.c: Fix error path of opening checkresult files (Andreas Ericsson)
[--snip--]
Previously we used to check that the returned filedescriptor
was larger than 0 to consider it a success and silently return
if it wasn't, but open() (as all other system calls) can return
0 on a successful invocation. Previously, this resulted in one
lost filedescriptor and the memory used for storing the path
to the checkresult file being leaked.

This patch takes care of both those leaks.
[--snip--]

patch is fine, removed one of those nifty printf hexcode
eastereggs too (as it's useless code).

fixes #423