base/nebmods.c: Replace local file-copy hack with my_fcopy() (Andreas Ericsson)
[--snip--]
my_fcopy() exists so we can copy files safely and have all the error
checking (and logging) done in one place. Implementing a local version
of it with less error-checking and logging is a bad idea.

While we're at it, we also make sure to release the output_file variable
if we fail to copy the file, plugging a small and largely insignificant
leak.
[--snip--]

fixes #424