cgi: add opt-in display_name from hosts/services definitions; if set, they will be shown in the classical ui
it is mentioned in the docs, that this feature will be done
somewhere in the future...

"The current CGIs do not use this option, although future
versions of the web interface will."

Well it is time for Icinga to do the first step on that.

If you set display_name in host definition, it will be shown
instead of host_name. for service definition, display_name will
be shown instead of service_description.

This is true for the following cgis:

* avail.cgi
* config.cgi
* extinfo.cgi
  - comments
  - downtime
  - scheduling queue
* histogram.cgi
* notifications.cgi
* outages.cgi
* status.cgi
  - host/service detail
  - host/service group overview
  - host/service problems
* statusmap.cgi
* statuswrl.cgi
* summary.cgi
* trends.cgi

display_name is also shown in the GD generated images.
If you don't set it, everything is like before.

fixes #307