rewrite icinga/ido2db initscript, not only depending on lockfile and removing it if start fails
1/ ido2db initscript was too dependant on the lockfile existing or not,
the routines now are the same as icinga core and it works flawlessly.
also if ido2db does not stop in a timely manner, the lockfile does not
get removed now.

2/ ido2db and icinga might fail after starting them. in this case the lockfile
and everything else needs to be removed. this is now done by checking on the status
and if no process running, showing an error and removing everything like the process
was stopped.

fixes #447
fixes #448