fix notification_period inheritance for services (Gordon Messmer)
The documentation for object inheritance indicates that where a
notification_period is not available for a service, the
notification_period from the associated host will be implicitly
inherited.

Right now, rather than inheriting any notification period,
each service has a "NULL" notification period which is treated
as a '24x7' period.  Now, Icinga will check for a NULL period
and use the period from the associated host as the documentation
states will happen.

fixes #456