idoutils: add config options to allow users decision on clearing realtime and config tables on Icinga core startup
Actual config objects are marked as (in)active, because
disabling the clearing results in historical config.

This feature was a users demand for keeping the
custom variables during Icinga restart, being more historical
and building own relationships on existing object_ids.

Existing and same config data will be updated, not added twice.

refs #469
refs #471