cgis: add config option show_all_services_host_is_authorized_for to cgi.cfg
It's possible to disable that the user can see all services only being
an authorized contact for the host.

If a user can only view the hosts and not the serviceshe's currently a
contact for, the default behaviour was that the user could view all
services only being authorized for the host.
But if it's demanded that the user must be authorized (a contact) for the
service too in order to view the service, change/add the new config option
set to 0.

was proposed back in 2006 ...

http://www.nagios-portal.org/wbb/index.php?page=Thread&postID=29655&highlight=cgiauth+patch#post29655

refs #357