error out if services defined without description in service object definition (service def and/or used templates)
previous fix bailed out when no service_description was found
in a service definition. In larger setups it could happen that
service descriptions are passed directly into a template where
a service is being inherited of.

This is rewritten into the service object resolving with all
associated templates. If no service_description set in the
end, verify config will fail with an error pointing to the
service definition file and linenumber, it mentions probable
template usages.

refs #451