rework lockfile handling in icinga/ido2db initscript
status check after startup has been dropped, does not really
work if icinga takes too long to fully startup.

instead before starting icinga/ido2db there is a check if
they are already running based on the lockfile (even if they
died already). so another check if the provided pid in the
lockfile is found - if not, this will be removed and icinga
/ido2db will be started normally.