remove catching a SIGSEGV on not dumping core and running as daemon
-* core: only catch SIGSEGV if we're not dumping core and running as a daemon (Andreas Ericsson)

applying this will cause idomod's 'error writing to data sink'
problems resulting in a caught SIGSEGV and shutting down the
core. Although it's an obvious problem of idomod removed for
the upcoming release.
It has to be resolved in idomod itsself afterwards.

Kudos to Andreas Bauer for reporting that.

fixes #531