first steps on postgresql with libpq
--enable-pgsql [--with-pgsql-inc|dir=]

configure checks for -lpq and libpq-fe.h and sets
the #defines HAVE_LIBPQ_FE_H and USE_PGSQL

code re-organized with #define USE_PGSQL, libdbi and
oracle work independant of my changes.

added initial pgsql database connection, which is tested
and working.

next to that, syslog output for ido2db added, if a client
is connecting.

refs #622