add opt-in log_external_commands_user in icinga.cfg, user will be logged by the core if enabled and sent through icinga.cmd
if an external command is sent with username before
commandargs, and if enabled in icinga.cfg, the given
username will be processed and logged.

the cgis read that config option and proceed sending
the correct syntax for that.

CMD;username;cmdargs instead of CMD;cmdargs

since this will break compatibility on existing log
(parsers), this is by default disabled, but might be useful
in the future.

the NEB API has not yet been update in order to support
the username being sent to command history.

refs #329