fix binding NULL to name1/2 in oracle; cleanup code; update icinga-api
the mysql patch in #363 introduced single quotes to binded values
in oracle. that's wrong for objects entries on name1/name2.
furthermore, NULL was binded as a string, which can't be the case
when selecting on IS NULL.

name1/2 must be explicitly binded to \0 and then NULL if occuring
NULL char*

next to that, the code is cleaned up a bit and the api submodule
pointer is uptodate for git submodule update

still, binding null values needs a rewrite - in it's own function.

refs #668
refs #622
refs #363
refs #468