fix oracle prep statement contactnotifmethoddata; add db trimming thread startup delay config option
contactnotificationmethoddata prepared statement was never
initialized, that's why the bindings failed, causing ido2db
to bail out. this is now fixed.

next to that, the db trimming startup delay can now be set in
ido2db.cfg - the default value is 60 seconds.
This might be useful to to extend the delay if the delete thread
might interfere at startup - next to that the trim_db_interval can
also be set as usual.

fixes #669
fixes #670