add eventhandler override (Sven Nierlein)
"Currently eventhandlers are executed in the main thread
[...] and block the core until the eventhandler is finished.
Attached is a patch, which makes it possible for eventhandling
neb modules to cancel/override the execution of eventhandlers.
This is currently only possible with service and host checks.
This patch would allow neb modules to stack eventhandlers on
external queues or even distribute execution of eventhandler.
This patch should not break the behavior of current neb modules,
because their return value is currently ignored and execution
of eventhandler will only be canceled on returning
"NEBERROR_CALLBACKOVERRIDE"."

refs #731