fix scheduled downtime notifications resent on restart/reload #584
What happened here is, that a downtime is rescheduled after restart.
So it reads the status from the retention file, but it has no idea
if the downtime is already started or not. Because the flag
"is_in_effect" isn't stored to the status or retention data.

Added the flag to status and retention data.
Also to every function which is adding a downtime.

When icinga gets started then it will register and trigger a downtime
event soon after start, bailing also out on sending a notification if
downtime->is_in_effect is TRUE.

done by Ricardo.

fixes #584