t-tap: Make test Makefile readable and sane (Andreas Ericsson)
We get rid of the ridiculously long lines and break things up a
little. We also utilize the Makefile auto-variable $^ when linking
and rely on the internal %.o: %.c buildrule to create the object
files for us.

In order to make this work, we have to sanitize test_checks so that
it doesn't #include the entire sourcefile it links to. Since one
still has to build Nagios (and the cgi's) before running the tests,
we can avoid double-building by linking to the build-product instead.

Signed-off-by: Andreas Ericsson <ae@op5.se>

Author: Andreas Ericsson <ae@op5.se>