core: fix problem with extra whitespace in command definition and ePN (Jason Young) #918
--SNIP--
If a command definition contains extra whitespace between the
filename and the first argument, this gets passed to the embedded
perl plugin as an empty string for the first argument. For example:

define command {
  command_name  test_check
  command_line  $USER1$/test_check.pl    $HOSTADDRESS$ $_HOSTSNMP$
                                      ^^^
}

will result in the plugin seeing $ARGV[0] as "".

This can cause the plugin to improperly interpret the arguments and
fail. Obviously the simple solution is to make sure your configuration
file does not have the extra spaces. However, this issue can be resolved
either in the p1.pl file or the c code. I have included patch files for
both options for your review. I am not an expert c programmer, so there
may be a better way to do this. The issue lies in the way
Text::ParseWords::parse_line handles leading whitespace. The shellwords
function performs the same task, but removes the leading whitespace.
--SNIP--

kudos to Jason Young.

refs #918