Serialize debug logging (Andreas Ericsson) #1035
--SNIP--
Since threaded modules can now utilize performance data and
checking APIs we must make sure that two processes never
write to the debug logging file at the same time. In truth,
stdio should take care of logging itself, but it seems as
if vfprinftf() doesn't handle that properly so we do it
ourselves, covering the case when one thread wants to move
the file and another wants to write to it.

In case child processes happen to inherit the lock (and thus
hang indefinitely while waiting for it), we use a soft locking
technique with a short loop that tries 5 times obtaining the
lock, with a sleep of 30 microseconds between each attempt. If
it fails to obtain it, it returns and ignores writing the
debug logging event. For obvious reasons, we can't log such
failed events, so they will pass silently.

Signed-off-by: Andreas Ericsson <ae@op5.se>
--SNIP--

refs #1035