threadsafaty: Use reentrant versions of localtime() and gmtime() (Andreas Ericsson) #1039
---
This makes Nagios/Icinga play well with Merlin if macros are exported
to environment variables.

If/when someone complains about Nagios not compiling on their
system, we'll introduce lock-wrapped versions of these functions
for use on those systems, but it's unlikely that a system that
supports threads doesn't have gmtime_r() and localtime_r() readily
available.

Reported-by: Patrik Båt <patrik.bat@cypoint.se>
Signed-off-by: Andreas Ericsson <ae@op5.se>
---

adapted for Icinga too, this will most likely attempt to fix the env var
problem with mk_livestatus too.

refs #1039