core: allow empty hostgroup aissgments for services as opt-in config option (Max Schubert) #949
--SNIP--
We used to bomb out with an error on such configurations, but with
this option it becomes optional if we should produce an error or a
warning.

It's useful to be able to assign services to empty hostgroups when
configuration files or pre-cached object files are distributed to
various pollers, or when the process of generating Nagios config is
automated, or when a set of services is slowly being phased out but
should be kept around in case of 'oh shit' epiphanies, such as one
might experience when phasing out old monitoring agents.

Thanks to Comcast in general and Max Schubert in particular for
sharing this patch.

Authored-by: Max Schubert <maxs@webwizarddesign.com>
Reviewed-by: Daniel Wittenberg <daniel.wittenberg.r0ko@statefarm.com>
Reviewed-by: Michael Friedrich <michael.friedrich@univie.ac.at>
Signed-off-by: Andreas Ericsson <ae@op5.se>
--SNIP--

this patch makes the checking for host_name on service definition
a warning in order to allow such behahvior.

fixes #949