classic ui: add trend timeperiod feature to locate the next problem (Carsten Emde) #1074
[--SNIP--]
In order to allow for scanning through the remaining recording and
searching for the occurrence of the next host or service problem,
the addtional time period tag TIMEPERIOD_NEXTPROBLEM was introduced.
If selected, the time period of the next diplay will be selected in
such a way that the full duration of the next problem will be
displayed at the center of the image. If only the start time is
available, because the problem is still present, the time window
of the image falls back to the 24-hour period.
[--SNIP--]

fixes #1074