idoutils: fix after core restart/reload on oracle db ORA-01403: no data found #1059
tracked down to unixts2date when the date calculation returns
no data for whatever reason. normally this should have been
added long time ago, but here it is ...

other exceptions will still raise, but no_data_found is just
returning a zero datetime string, so that the application
will have to figure out theirselves what to do in that case.

idoutils will just insert the zero value, and let the application
decide then.

thx to Ronny for the ideas on fixing it.

fixes #1059