core: save processed command for neb callback #1185
this introduces a new member in host|service struct
in objects.h but as a matter of fact, it'll be
added at the end, non-violating typecasts and faulty
address mapping. Kudos to Opsview for the patches,
we'll take the more processing at the end of the struct
(which is interessting regarding performance vs
compatibility).

tested with IDOUtils and MKLivestatus (1.1.9i8)

fixes #1185