Add an optional alternative CGI driven view for the top frame #1406
This enables the ability to display CGI content in the 'top' frame of
the Classic UI that has a similar feel to the top header in icinga-web
including tactical and monitor performance information.

index.html now calls tac.cgi?tac_header instead of top.html for the
content of the top frame. By default the output essentially looks
the same as what was in top.html.

The new CGI top view can be enabled with the cgi.cfg option:

show_tac_header=1

refs #1406