core: add 'module' as object config, allowing cfg_dir usage loading multiple modules without touching broker_module in icinga.cfg #162
define module{
        module_name     idomod
        module_type     neb
        path            /usr/local/icinga/bin/idomod.o
        args            config_file=/usr/local/icinga/etc/idomod.cfg
}

this is *fully* optional to the default way of adding event broker modules.

it just shows an initial proof of concept which steps are needed to define
new objects and their basic usage.

and by adding this, every packager can ignore icinga.cfg:broker_module
by just installing idoutils.cfg into etc/modules/ or similar, just including
the cfg_dir and allowing this object config to be read, and the event broker
modules to be loaded.

pls read the issue for further information on that
https://dev.icinga.org/issues/162

refs #162