Additional corrections to tac header counts #1508
This corrects the corrections made by rbartels to the tac header
counts in commit 8bc877224fb82a78f8f8e9ca43ec3805b2c7ca58

It adds additional service_*_unacknowledged_hosts_down ints and
takes them into account in the display of the service counts.

It also makes the display of pending states that Ricardo included
with his chages in the tac header a cgi.cfg option that is off by
default.

The count entries in the tac header begin to lose room and may wrap
for users if they have a display that is <1024x768 when the pending
entries are included. That was the primary reason why I didn't
include them originally. However, since Ricardo made the effort
to add them and those with larger screens may find them useful, I
felt this would be a reasonable compromise.

refs #1508
refs #1506
refs #1505