Added option to suppress coloring of hosts/services in maintenance #992
The current Classic UI view shows the state of hosts including
attention grabbing colors even if the host/service is scheduled for
downtime. This tends to distract from discerning which problems
really exist between those which were planned or expected.

This commit adds the following option to cgi.cfg:

suppress_maintenance_downtime=0

It is disabled by default to maintain the old behavior, however
those wishing to enable it will note that rather than displaying
issues in their regular color, they will be grey (similar to
the pending state). The status, however will still display the
actual state and this is intended behavior. This is so the user
can use icinga to see what the state is in before the downtime
ends. A useful example for this is so it can be seen that a
host/service that was being worked on actually came back online.

In addition, hosts in scheduled downtime will be grey in the
statusmap.cgi

refs #992