Added option to show partial hostgroups #1013
By default, a user only sees a hostgroup and the hosts within it if
they are an authorized contact for all of the hosts of the group.

This commit adds the cgi.cfg option:

show_partial_hostgroups=0

When enabled, the hostgroups overview will show a partial listing
of hosts that the user is an authorized contact for within each
hostgroup.

It will also add the string "(Partial Hostgroups Enabled)" to the
top of the Hostgroup Overview to help prevent any confusion over
whether the option is in use or not. However for privacy reasons,
hostgroups that are only showing a partial listing are not
specifically indicated.

COMPATABILITY NOTICE: As with any tweak made to the output of the
CGIs, enabling this option may adversely impact third party
programs that rely on 'screen scraping' to get their information.
If you encounter such a problem, turn this option back to it's
default of off and encourage the developer(s) of the program to
use JSON for their data needs instead.

refs #1013