core: avoid sending notifications to the wrong contact (Andreas Ericsson) #1654
if a neb module blocks notifications due to recent changes
on the source, the notification_list wasn't cleared, causing
the next notification to hit all contacts for the blocked
notification as well as to all contacts for the current
notification.

clearing it on the neb module returns, and also prior to
filling it, as an empty NULLed least is not affecting
performance.

refs #1654