Prevent empty partial hostgroups from displaying
The initial change to status.c for partial hostgroup display made it
so that users would see empty hostgroups even for the hostgroups they
are not authorized to see.

This change prevents that from occuring by pre-iterating over the
hostgroup and suppressing the display of it if the hostgroup would
be empty.

refs #1013