idoutils: add threaded circular buffer to ido2db, same as idomod #1259
main problem - the overall dump is now not synchronous but asynchronous.
this requires finetuning on both sides, but the overall core blocking
seems to be removed all over.

furthermore, making all functions thread safe will be needed.
and ido2db code needs to be cleaned from the old attempts.

and, there souldn't be retries on the buffer if full anymore,
the 2 buffers should just get into blocking mode.

refs #1259