Change markup color for unreachable hosts in statusmap.c #1464
This commit marks hosts that are UNREACHABLE with pink rather than
with red in statusmap.c.

This makes the markup color logically consistent with the other CGIs
and gives a bit more accurate information within the status map
regarding what the issue is with a particular host(s). It also allows
for easier identification of the hosts that are the root cause of an
issue.

The background coloring remains red for visual continuity since
UNREACHABLE is still a problem of sorts.

refs #1464