classic ui: fixed search not returning all results, only first match (without * at the end) #844
refs: #844

Done a rewrite of the search part. If user is searching something and a host or service
matches, then the status entry gets marked as a match. When statusdata gets displayed
it only shows the matching one's. Added "search_matched" to both statusdata struct.

The filter uses regular expressions. Following items get checked against search string:

* Host
** Host Name
** Display Name
** Alias
** Address
** Address6

* Service
** Description
** Display Name
** Host Name
** combination of hostname and service description "hostname servicedescription"

If all of this fails it tries to find a hostgroup or servicegroup which has to match the
exact search string. This is still the old behavior.