core: fix event removal from queues with O(1) removal from doubly linked lists (Andreas Ericsson) #2183
given the fact that the algorithm before was violating the purpose
of a double linked list, looping over the event list if the on top
event was not found in the first place, this is now revamped.

since the event itsself provides next and prev pointers already,
we can work with the "next" and the "previous" events in the list.

if there is a previous event before the current one, we modify its
next pointer to target the events' next event, leaving out the event
to remove.
same goes for the next event, if there, its previous pointer leaves
out the event to remove, but targetting the event before.

in case there is no previous event, the head of the list is modified
to point to the next event after the removed one.
same matches for the tail of the list, assigning the previous and now
last event of the list.

if there was only one element in the list, the last 2 condition would
match, setting the head and tail pointers to NULL, meaning to say -
1 list, removed 1 element, now empty list.

this algorithm allows to remove an event just by reordering next and
previous pointers within a doubly linked list, and preventing us to loop.

so to say, Andreas is a genius ...

refs #2183