core: fix handling of allow_empty_hostgroup_assignment when hitting NULL'edhost_name attribute, now properly finishing #949
previously we were checking if host_name attribute was NULL, warning the
user about that and then looping to the objects_skiplist_insert and
finally erroring out due to wrongly to-be-registered incomplete object.
this is now fixed for both, the default is an error, using
allow_empty_hostgroup_assignment will now discard the warning and
ignoring the object.

oh, and updated the config description a bit, to tell where this
functionality is useful (and somewhat not to do some object tricks like
people tried to do and bug report then...)

refs #949