fix test_checks in t-tap, strange sigsegv all the way ...
passing a checkresult from services straight to a host
without checking that the svcdesc is null'ed leads to
rather strange behaviour. furthermore, using calloc
instead of malloc does not prevent us from initializing
all the important members which reach the inner core
functionality already set - we shouldn't fail on the
tests just because we don't know how to set variables.

refs #1192