idoutils: fix pgsql unix_timestamp cannot handle timstamp with timezone #2203
as a matter of fact, the columns were changed to
understand timestamp with timezone, but not the
conversion functions, still ignoring the timezone.

this patch fixes that accordingly, replacing
from_unixtime and unix_timestamp with timezone
aware functionality.

this possibly affects the sla and tackle problems
with postgresql in icinga-web, as well as other
instance cronk time calculations too (#2293)

refs #2203

refs #2393